Добавлен параметр компиляции -Werror Заменены очереди на буферные потоки. Не проверено.

+++ xmobile-fw/FreeRTOS/Source/portable/GCC/ATMega2560/port.c
@@ -214,7 +214,7 @@ uint16_t usAddress;
 	pxTopOfStack--;
 	*/
 
-	usAddress = pxCode;
+	usAddress = (uint16_t)pxCode;
 	*pxTopOfStack = ( usAddress & 0x00ff );
 	pxTopOfStack--;
